Bump 2.29.1: fix CLI crash on top-level require('electron')
src/main/search/match.ts started importing splitContent from src/main/ipc/logs.ts in v2.27.0. ipc/logs.ts top-level-imports 'electron' for ipcMain, which esbuild keeps external in the CLI bundle — so every condash invocation crashed at startup with MODULE_NOT_FOUND under ELECTRON_RUN_AS_NODE=1, taking /projects, /knowledge, /skills, /tidy with it.

Extract the pure parsing helpers (META_LINE_PREFIX, parseMetaLine, splitContent, HeaderJson, FooterJson) into src/main/logs-format.ts — CLI-safe by construction, no electron and no @xterm/headless imports. terminal-logger.ts and ipc/logs.ts now consume them from there; match.ts imports splitContent directly from logs-format. ipc/logs.ts re-exports splitContent for back-compat. Same pattern as the user-data-dir.ts extraction from settings.ts.
