+++ CLAUDE.md
@@ -54,7 +54,7 @@ When changing the dev port, update **every** file:
 
 - **Per-machine**: `${XDG_CONFIG_HOME:-~/.config}/condash/settings.json`. Owned by Electron's `app.getPath('userData')`, derived from `package.json`'s `name` field. JSON only. Sole owner of `lastConceptionPath` + `recentConceptionPaths` (cap 5); also carries global defaults for every workspace key.
 - **Per-conception**: `<conception>/.condash/settings.json` (canonical). The whole `.condash/` directory is fully gitignored by default — settings included — so it carries per-host state with no commit-leak risk. Two legacy paths still read as fallbacks: `<conception>/condash.json` (old canonical) and `<conception>/configuration.json` (legacy²). Writes always target `.condash/settings.json`. The auto-migrator (`src/main/condash-dir-migrate.ts`) copies legacy content on first open, tombstones the source file, and appends `.condash/` to `.gitignore` when the conception is a git repo. `condash config migrate` exposes the same operation explicitly.
-- **Per-conception logs**: `<conception>/.condash/logs/YYYY/MM/DD/HHMMSS-<sid>.txt`. **One single plain-text file per pty spawn** since v2.27.0 — no sidecar, no compression. Metadata travels inside the `.txt` as two `# condash: {...}` JSON lines: a header line at the top (always present, carries `{sid, side, repo?, cwd, cmd, argv, started}`) and a footer line at the bottom (only after `exit()`, carries `{finished, exitCode}`). `cat foo.txt` shows the metadata too. Gitignored alongside `settings.json`. **Opt-in** since 2.25.0 — `terminal.logging.enabled` defaults to `false`; flip it in `Settings → Terminal → Logging` to start recording. The pref is read once at session-start (writer construction). The writer (`src/main/terminal-logger.ts`) pipes pty bytes into a headless xterm (`@xterm/headless`), every 5 s renders the buffer row-by-row via `IBufferLine.translateToString(true)` (plain text — no SGR, no CSI, no cursor-forward), composes header + body (+ footer if exited), and atomically rewrites the `.txt`. Flushes are serialised through `flushChain` so `close()` can await every pending write before disposing the xterm. `in` keystrokes are *not* captured — the pty echoes them back through `out`. Session size is bounded by xterm scrollback (`terminal.logging.scrollback`, default 10000 lines); no rotation. Janitor (`src/main/terminal-logger-janitor.ts`) (1) evicts day-directories past `terminal.logging.retentionDays` (default 14), (2) evicts oldest-first while over `terminal.logging.maxDirMb` (default 500); runs at startup + every 24 h. The Logs working surface (`Cmd+Shift+L`) is now a **sessions list grouped by day** (`src/renderer/panes/logs.tsx`) — each row is a card showing time · repo · cmd · exit · size. Clicking a card opens `LogsViewerModal` (`src/renderer/logs-modal.tsx`): a full-overlay viewer with a virtualised plain-text transcript (fixed-row-height windowing over `lines.length * --logs-row-height` via an absolutely-positioned slice), a search box that pre-computes per-line hit indices and scrolls the active hit into view, and `Esc` to close. Long lines horizontal-scroll (no wrap) so the virtualizer's fixed-row-height math holds. Logs are also a source in the global search modal — `src/main/search/walk.ts:collectLogFiles` walks `<root>/YYYY/MM/DD/*.txt`; `src/main/search/match.ts` strips the `# condash:` header / footer lines via `splitContent` (exported from `ipc/logs.ts`) before substring matching, so snippets don't quote the JSON. Activating a log hit posts a `LogsOpenRequest` that opens the modal directly. Legacy `.jsonl` files from condash ≤ 2.22 and `.txt.gz` from 2.23–2.26 are ignored by the viewer — the janitor evicts them by age.
+- **Per-conception logs**: `<conception>/.condash/logs/YYYY/MM/DD/HHMMSS-<sid>.txt`. **One single plain-text file per pty spawn** since v2.27.0 — no sidecar, no compression. Metadata travels inside the `.txt` as two `# condash: {...}` JSON lines: a header line at the top (always present, carries `{sid, side, repo?, cwd, cmd, argv, started}`) and a footer line at the bottom (only after `exit()`, carries `{finished, exitCode}`). `cat foo.txt` shows the metadata too. Gitignored alongside `settings.json`. **Opt-in** since 2.25.0 — `terminal.logging.enabled` defaults to `false`; flip it in `Settings → Terminal → Logging` to start recording. The pref is read once at session-start (writer construction). The writer (`src/main/terminal-logger.ts`) pipes pty bytes into a headless xterm (`@xterm/headless`), every 5 s renders the buffer row-by-row via `IBufferLine.translateToString(true)` (plain text — no SGR, no CSI, no cursor-forward), composes header + body (+ footer if exited), and atomically rewrites the `.txt`. Flushes are serialised through `flushChain` so `close()` can await every pending write before disposing the xterm. `in` keystrokes are *not* captured — the pty echoes them back through `out`. Session size is bounded by xterm scrollback (`terminal.logging.scrollback`, default 10000 lines); no rotation. Janitor (`src/main/terminal-logger-janitor.ts`) (1) evicts day-directories past `terminal.logging.retentionDays` (default 14), (2) evicts oldest-first while over `terminal.logging.maxDirMb` (default 500); runs at startup + every 24 h. The Logs working surface (`Cmd+Shift+L`) is now a **sessions list grouped by day** (`src/renderer/panes/logs.tsx`) — each row is a card showing time · repo · cmd · exit · size. Clicking a card opens `LogsViewerModal` (`src/renderer/logs-modal.tsx`): a full-overlay viewer with a virtualised plain-text transcript (fixed-row-height windowing over `lines.length * --logs-row-height` via an absolutely-positioned slice), a search box that pre-computes per-line hit indices and scrolls the active hit into view, and `Esc` to close. Long lines horizontal-scroll (no wrap) so the virtualizer's fixed-row-height math holds. Logs are also a source in the global search modal — `src/main/search/walk.ts:collectLogFiles` walks `<root>/YYYY/MM/DD/*.txt`; `src/main/search/match.ts` strips the `# condash:` header / footer lines via `splitContent` (exported from `src/main/logs-format.ts` — the pure-parsing module shared between the renderer-side `ipc/logs.ts`, the writer in `terminal-logger.ts`, and the CLI's search command; carved out so the CLI bundle's import graph doesn't touch `electron` or `@xterm/headless`) before substring matching, so snippets don't quote the JSON. Activating a log hit posts a `LogsOpenRequest` that opens the modal directly. Legacy `.jsonl` files from condash ≤ 2.22 and `.txt.gz` from 2.23–2.26 are ignored by the viewer — the janitor evicts them by age.
 - `CONDASH_CONCEPTION_PATH` env var override: wired in both Electron (`src/main/settings.ts`) and CLI (`src/cli/conception.ts`). See [`docs/guides/configure-conception-path.md`](docs/guides/configure-conception-path.md) for the resolution chain. The conception-detector accepts any of the three recognised config files when validating a candidate path.
 - The merge resolver lives in `src/main/effective-config.ts`. Every reader (`repos.ts`, `worktree-ops.ts`, `launchers.ts`, `audit.ts`, `terminals.ts`, `conception-paths.ts`, the CLI's `config` verbs) goes through it.
 - **Merge rule**: top-level replace for every key, with **one exception**: `terminal` merges one level deep so a conception customising `terminal.logging` keeps the per-machine `terminal.launchers` / `screenshot_dir` / shortcuts. The exception is documented in `docs/reference/config.md` and tested in `effective-config.test.ts`.
